fix: heartbeat function place
